seperate out migration total component so rerenders less disruptive
