+++ src/modules/forking/components/migrate-rep-form/migrate-rep-form.jsx
@@ -10,6 +10,7 @@ import { SCALAR } from 'modules/markets/constants/market-types'
 import { ExclamationCircle as InputErrorIcon } from 'modules/common/components/icons'
 import selectMigrateTotals from 'modules/reporting/selectors/select-migrated-totals'
 import Styles from 'modules/forking/components/migrate-rep-form/migrate-rep-form.styles'
+import FormattedMigrationTotals from 'modules/forking/components/migrate-rep-form/formatted-migration-totals'
 
 export default class MigrateRepForm extends Component {
 
@@ -33,36 +34,10 @@ export default class MigrateRepForm extends Component {
       inputRepAmount: '',
       inputSelectedOutcome: '',
       scalarInputChoosen: false,
-      formattedMigrationTotals: null,
-      blockNumber: props.currentBlockNumber,
     }
 
     this.focusTextInput = this.focusTextInput.bind(this)
-  }
-
-  componentWillMount() {
-    this.getForkMigrationTotals()
-  }
-
-  componentWillReceiveProps(newProps) {
-    const updateBlock = createBigNumber(this.state.blockNumber)
-    const currentBlock = createBigNumber(newProps.currentBlockNumber)
-    if (currentBlock.gt(updateBlock)) {
-      this.setState({
-        blockNumber: this.props.currentBlockNumber,
-      })
-    }
-  }
-
-  shouldComponentUpdate(nextProps, nextState) {
-    const updateBlock = createBigNumber(this.state.blockNumber)
-    const currentBlock = createBigNumber(nextProps.currentBlockNumber)
-    if (currentBlock.gt(updateBlock) || this.props.accountREP !== nextProps.accountREP) return true
-    return false
-  }
-
-  componentWillUpdate() {
-    this.getForkMigrationTotals()
+    this.validateOutcome = this.validateOutcome.bind(this)
   }
 
   componentDidUpdate(prevProps) {
@@ -85,20 +60,6 @@ export default class MigrateRepForm extends Component {
       })
     }
 
-  }
-
-  getForkMigrationTotals() {
-    const {
-      getForkMigrationTotals,
-    } = this.props
-    getForkMigrationTotals((err, forkMigrationTotals) => {
-      if (err) return console.error(err)
-      const { reportableOutcomes } = this.props.market
-      const formattedMigrationTotals = selectMigrateTotals(reportableOutcomes, forkMigrationTotals)
-      this.setState({
-        formattedMigrationTotals,
-      })
-    })
   }
 
   checkRepAmount(repAmount, updatedValidations) {
@@ -138,9 +99,10 @@ export default class MigrateRepForm extends Component {
     })
   }
 
-  validateOutcome(validations, selectedOutcome, selectedOutcomeName, isMarketInValid) {
+  validateOutcome(selectedOutcome, selectedOutcomeName, isMarketInValid) {
     const {
       updateState,
+      validations
     } = this.props
     const updatedValidations = { ...validations }
     updatedValidations.selectedOutcome = true
@@ -232,6 +194,7 @@ export default class MigrateRepForm extends Component {
       selectedOutcome,
       selectedOutcomeName,
       validations,
+      getForkMigrationTotals,
     } = this.props
 
     const {
@@ -239,6 +202,7 @@ export default class MigrateRepForm extends Component {
       inputSelectedOutcome,
       inputRepAmount,
     } = this.state
+
     return (
       <ul className={classNames(Styles.MigrateRepForm__fields, FormStyles.Form__fields)}>
         <li>
@@ -251,20 +215,13 @@ export default class MigrateRepForm extends Component {
                 <span>Outcome</span>
               </label>
             </li>
-            { formattedMigrationTotals && formattedMigrationTotals.length > 0 && (formattedMigrationTotals).map(outcome => (
-              <li key={outcome.id}>
-                <button
-                  className={classNames({ [`${FormStyles.active}`]: selectedOutcome === outcome.id })}
-                  onClick={(e) => { this.validateOutcome(validations, outcome.id, outcome.name, false) }}
-                >{outcome.name === 'Indeterminate' ? 'Market is Invalid': outcome.name}
-                  <span className={Styles.MigrateRepForm__outcome_rep_total}>{ (outcome && outcome.rep.formatted) || '0'} REP Migrated</span>
-                  { outcome && outcome.winner &&
-                    <span className={Styles.MigrateRepForm__winning_outcome}> WINNING UNIVERSE</span>
-                  }
-                </button>
-              </li>
-            ))
-            }
+            <FormattedMigrationTotals 
+              validateOutcome={this.validateOutcome} 
+              market={market} 
+              getForkMigrationTotals={getForkMigrationTotals} 
+              currentBlockNumber={this.props.currentBlockNumber} 
+              selectedOutcome={selectedOutcome} 
+            />
             { market.marketType === SCALAR &&
               <ul className={FormStyles['Form__radio-buttons--per-line-long']}>
                 <li>
